Visual Representation Almost Complete
I am almost finished coding the visual representation, which looks very close to the original design.

There's just a few tweaks I need to do, and once that's done I can then begin coding JS code and integrating it with the button, so that the visual representation of the grid will change on submit.