Finally finished with breakpoints
At this point, with the website being tested on my laptop, as well as on PC using 000webhost, I think it looks very close to the original design.

Only thing I would improve in the later future, would be adjusting padding from top and bottom, because it's not equal as the padding from the sides.

But, anyway, now I will work with JS. I will implement a simple script that hides a section of code and displays another one instead.