Final update
Finished my wonky JS file, it work as it should, it hides the grid and displays another page when submitted. It refreshes when dismiss button is clicked.

When mouse moves out of the email box, it highlights in red and error text shows up.

Responsive in mobile, tablet and pc.

Not perfectly identical to the design, but i'm proud.