Website Finished
While it does not look 100% identical, I am very proud of this website.

It took a little bit longer, but in the end I am satisfied with the result.

The form is fully functional, box highlights red when mouse moves out of it, and goes back to normal when something is entered.

When you try to submit an empty form, an alert will pop out with instructions.

Upon successful submission, another page pops up, with the button to dismiss which refreshes the page.

In overall, I am satisfied, result not prefect as to the design, but it works and it looks very similar.